+++ src/model-provider/llamacpp/LlamaCppTextEmbeddingModel.ts
