+++ src/model-provider/llamacpp/index.ts
@@ -1,4 +1,4 @@
 export { LlamaCppError, LlamaCppErrorData } from "./LlamaCppError.js";
-export * from "./LlamaCppEmbeddingModel.js";
+export * from "./LlamaCppTextEmbeddingModel.js";
 export * from "./LlamaCppTextGenerationModel.js";
 export * from "./LlamaCppTokenizer.js";
